# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KotobaCore API",
    "description": "FastAPI application for KotobaCore (FR-092).\n\nEndpoints\n    GET  /health              liveness\n    GET  /version             KotobaCore / schema / API versions\n    POST /analyze             {text, document?, reference_date?} → Semantic IR (AnalysisResult JSON)\n    POST /query               {text, reference_date?}            → Query IR\n    POST /chunk               {text, max_chars?, min_chars?}     → document chunks\n    POST /tokenize            {text, granularity?}               → tokens\n\nSecurity (§10): optional bearer-token auth (``api_token`` / KOTOBACORE_API_TOKEN),\noptional per-client rate limit (``rate_limit`` requests per minute /\nKOTOBACORE_RATE_LIMIT), a request size cap (``max_chars``), and request bodies\nare never logged. Errors follow §9: ``{\"error\": {\"code\": \"E7xx …\", \"message\": …}}``.\n\nThe analyzer is built once per app; a separate process from the Streamlit demo.\n",
    "version": "1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/version": {
      "get": {
        "summary": "Version",
        "operationId": "version_version_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Version Version Get"
                }
              }
            }
          }
        }
      }
    },
    "/analyze": {
      "post": {
        "summary": "Analyze",
        "operationId": "analyze_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Analyze Analyze Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "post": {
        "summary": "Query",
        "operationId": "query_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Query Query Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chunk": {
      "post": {
        "summary": "Chunk",
        "operationId": "chunk_chunk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Chunk Chunk Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tokenize": {
      "post": {
        "summary": "Tokenize",
        "operationId": "tokenize_tokenize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Tokenize Tokenize Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Japanese text"
          },
          "document": {
            "type": "boolean",
            "title": "Document",
            "description": "Multi-paragraph document → paragraphs / sentences / document_chunks",
            "default": false
          },
          "reference_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Date",
            "description": "YYYY-MM-DD for relative dates (今日 / 来月)"
          },
          "semantic_only": {
            "type": "boolean",
            "title": "Semantic Only",
            "description": "Drop tokens / semantic_tokens from the response",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AnalyzeRequest"
      },
      "ChunkRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "max_chars": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 40.0,
            "title": "Max Chars",
            "default": 400
          },
          "min_chars": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 0.0,
            "title": "Min Chars",
            "default": 80
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ChunkRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "QueryRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "reference_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Date"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "QueryRequest"
      },
      "TokenizeRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "granularity": {
            "type": "string",
            "pattern": "^(coarse|fine)$",
            "title": "Granularity",
            "default": "coarse"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TokenizeRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}